requirements.txt
python-telegram-bot[job-queue]==22.7
httpx>=0.27,<0.29
aiohttp==3.13.5
python-dotenv==1.2.2
solders>=0.21.0
cryptography>=42.0.0
tweepy>=4.14.0
redis>=5.0.0
apscheduler>=3.10.0
google-generativeai>=0.8.0
google-auth>=2.20.0
Pillow>=10.0.0
